fix symbol id lookup
