+++ packages/sdks/src/functions/get-content/types.ts
@@ -1,0 +1,21 @@
+export interface GetContentOptions {
+  /** The model to get content for */
+  model: string;
+  /** Your public API key */
+  apiKey: string;
+  /** A/B test groups the current visitor is in. Key is the content ID, value is the variation ID */
+  testGroups?: Record<string, string> | null;
+  /** Number of items to fetch. Default is 1 */
+  limit?: number;
+  /** User attributes to target on, such as { urlPath: '/foo', device: 'mobile', ...etc } */
+  userAttributes?: (Record<string, string> & { urlPath?: string }) | null;
+  /** Custom query */
+  query?: Record<string, any>;
+  /** Other API options as key:value pairs */
+  options?: Record<string, any>;
+  /**
+   * If set to `true`, it will lazy load symbols/references.
+   * If set to `false`, it will render the entire content tree eagerly.
+   */
+  noTraverse?: boolean;
+}
